requirements.txt
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.3.0
lightgbm>=4.1.0
tensorflow>=2.15.0
keras>=3.0.0
yfinance>=0.2.0
TA-Lib>=0.4.28
PyYAML>=6.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
requests>=2.31.0
matplotlib>=3.8.0
seaborn>=0.13.0
ipykernel
